Save survey URL without worker ID on the end.
Previously, the worker ID was passed to gateway.php twice: once as part of the survey URL and again as a separate parameter. I have simplified this by passing along a "passId" parameter through the URL instead, which tells gateway.php whether to pass along the worker ID or not.